#81, Refactored host measurements, as well as clean up of multiple issues around refactoring.
Added tests.

Moved measurements objects to specs.py in order to be able to avoid circular reference.
